divide materials from element_commands
